Function

SP2000China.GetFamiliesByFamilyName — Function

GetFamiliesByFamilyName(;family_name::String,api_key::String,page::Int=1)

Description

Query by family name and return a collection of family IDs.

Parameters

family_name: The family name, or a part of it, supporting both Latin and Chinese names.

api_key: The API service key for registered users.

page: The page number, an integer not less than 1. If not provided, it defaults to 1.

Results

result: A structure of type GetFamiliesByFamilyNameStruct.

result.data: Dictionary converted from JSON information.

result.count: Total number of matches.

result.page: Current data page.

reult.limit: Number of items displayed per page.

result.abstract: Refined data frame.

Example

using SP2000China;
your_family_name = "Coriariaceae";
your_api_key = "Please register an account and obtain an API key";
your_page = 1;
result = GetFamiliesByFamilyName(family_name=your_family_name,api_key=your_api_key,page=your_page);
result.data
result.count
result.page
reult.limit
result.abstract

SP2000China.GetSpeciesByFamilyId — Function

GetSpeciesByFamilyId(;family_id::String,api_key::String,page::Int=1)

Description

Query species by family ID and return a list of species.

Parameters

family_id:The family ID, a unique value.

api_key:The API service key for registered users.

page:The page number, an integer not less than 1. Defaults to 1 if not provided.

Results

result: A structure of type GetSpeciesByFamilyIdStruct.

result.data: Dictionary converted from JSON information.

result.count: Total number of matches.

result.page: Current data page.

result.limit: Number of items displayed per page.

result.abstract: Refined data frame.

Example

using SP2000China;
your_family_id = "F20171000000291";
your_api_key = "Please register an account and obtain an API key";
your_page = 1;
result = GetSpeciesByFamilyId(family_id=your_family_id,api_key=your_api_key,page=your_page);
result.data
result.count
result.page
reult.limit
result.abstract

SP2000China.GetSpeciesByScientificName — Function

GetSpeciesByScientificName(;scientific_name::String,api_key::String,page::Int=1)

Description

Query by scientific name and return species IDs.

Parameters

scientific_name:The scientific name, or a part of it, supporting both Latin and Chinese names.

api_key: The API service key for registered users.

page: The page number, an integer not less than 1. If not provided, it defaults to 1.

Results

result: A structure of type GetSpeciesByScientificNameStruct.

result.data: Dictionary converted from JSON information.

result.count: Total number of matches.

result.page: Current data page.

reult.limit: Number of items displayed per page.

result.abstract: Refined data frame.

Example

using SP2000China;
your_scientific_name = "Actinidia arg";
your_api_key = "Please register an account and obtain an API key";
your_page = 1;
result = GetSpeciesByScientificName(scientific_name=your_scientific_name,api_key=your_api_key,page=your_page);
result.data
result.count
result.page
reult.limit
result.abstract

SP2000China.GetSpeciesByCommonName — Function

GetSpeciesByCommonName(;common_name::String,api_key::String,page::Int=1)

Description

Query by common name and return species or subspecies information.

Parameters

common_name: The common name, or a part of it.

api_key: The API service key for registered users.

page: The page number, an integer not less than 1. If not provided, it defaults to 1.

Results

result: A structure of type GetSpeciesByCommonNameStruct.

result.data: Dictionary converted from JSON information.

result.count: Total number of matches.

result.page: Current data page.

reult.limit: Number of items displayed per page.

result.abstract: Refined data frame.

Example

using SP2000China;
your_common_name = "土人参";
your_api_key = "Please register an account and obtain an API key";
your_page = 1;
result = GetSpeciesByCommonName(common_name=your_common_name,api_key=your_api_key,page=your_page);
result.data
result.count
result.page
reult.limit
result.abstract

SP2000China.GetSpeciesByNameCode — Function

GetSpeciesByNameCode(;name_code::String,api_key::String)

Description

Retrieve detailed information based on the species ID.

Parameters

name_code: The species ID.

api_key: The API service key for registered users.

Results

result: A structure of type GetSpeciesByNameCodeStruct.

result.data: Dictionary converted from JSON information.

result.abstract: Refined data frame.

Example

using SP2000China;
your_name_code = "1ac19d0d82d84dd2900d51a742fa9296";
your_api_key = "Please register an account and obtain an API key";
result = GetSpeciesByNameCode(name_code=your_name_code,api_key=your_api_key);
result.data
result.abstract

SP2000China.GetNameByKeyword — Function

GetNameByKeyword(;keyword::String,api_key::String,page::Int=1)

Description

Search for name information based on a keyword.

Parameters

keyword: The name keyword (at least 2 characters).

api_key: The API service key for registered users.

page: The page number, an integer not less than 1. If not provided, it defaults to 1.

Results

result: A structure of type GetNameByKeywordStruct.

result.data: Dictionary converted from JSON information.

result.count: Total number of matches.

result.page: Current data page.

reult.limit: Number of items displayed per page.

result.abstract: Refined data frame.

Example

using SP2000China;
your_keyword="柳莺";
your_api_key = "Please register an account and obtain an API key";
your_page = 1;
result = GetNameByKeyword(keyword=your_keyword,api_key=your_api_key,page=your_page);
result.data
result.count
result.page
reult.limit
result.abstract

Other

SP2000China.SP2000China — Module

The Julia package for obtaining information on the list of biological species, SP2000China.

SP2000China.GetFamiliesByFamilyNameStruct — Type

Construct a structure to store the results of the function GetFamiliesByFamilyName.

SP2000China.GetSpeciesByFamilyIdStruct — Type

Construct a structure to store the results of the function GetSpeciesByFamilyId.

SP2000China.GetSpeciesByScientificNameStruct — Type

Construct a structure to store the results of the function GetSpeciesByScientificName.

SP2000China.GetSpeciesByCommonNameStruct — Type

Construct a structure to store the results of the function GetSpeciesByCommonName.

SP2000China.GetSpeciesByNameCodeStruct — Type

Construct a structure to store the results of the function GetSpeciesByNameCode.

SP2000China.GetNameByKeywordStruct — Type

Construct a structure to store the results of the function GetNameByKeyword.